+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 0.0.6 (2026-01-18)
+1. improve hints in situation of no git users config
+2. remove truncate in list output
+
+
 # Version 0.0.5 (2025-12-27)
 
 update wxs file